fix linux bootstrap build after previous changes

+++ bld/cg/c/dumpopt.c
@@ -113,7 +113,7 @@ static  bool    LblName( code_lbl *lbl ) {
     if( lbl->lbl.sym == NULL ) return( TRUE );
     DumpLiteral( "(" );
     if( AskIfRTLabel( lbl ) ) {
-        DumpXString( AskRTName( (rt_class)(uintptr_t)lbl->lbl.sym ) );
+        DumpXString( AskRTName( (rt_class)(int)lbl->lbl.sym ) );
     } else if( AskIfCommonLabel( lbl ) ) {
         DumpLiteral( "Common import => [" );
         DumpInt( (int)lbl->lbl.sym );
